refactor: store를 각 feature 경로로 이동
